misc: clean up and ffmpeg record helper
