+++ Assets/Scripts/Misc/Types/Enums/RecordMode.cs
@@ -4,5 +4,6 @@
     {
         Disable,
         OBS,
+        FFmpeg
     }
 }
