+++ Assets/Scripts/Utils/FFmpegRecordHelper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f2c64ad9ad4eee94eb2fae3673db236e
+guid: 79b28cd4eea5e7548b4a50dad61fecd2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
